Fill out quiz aanpassingen

+++ src/main/java/controller/ManageQuestionsController.java
@@ -8,12 +8,14 @@ import database.mysql.QuestionDAO;
 import database.mysql.QuizDAO;
 import javafx.fxml.FXML;
 import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.ListView;
 import javafx.scene.control.TextField;
 import model.Question;
 import view.Main;
 
 import java.util.ArrayList;
+import java.util.Optional;
 
 public class ManageQuestionsController {
 
@@ -27,30 +29,28 @@ public class ManageQuestionsController {
 
     public ManageQuestionsController () {
         this.dbAccess = Main.getDBaccess();
+        this.questionDAO = new QuestionDAO(dbAccess);
     }
 
     public void setup() {
-        this.questionDAO = new QuestionDAO(this.dbAccess);
         ArrayList<Question> alleVragen = questionDAO.getAllperLoggedInCoordinator(Main.loggedInUser.getIdUser());
         for (Question question: alleVragen) {
             questionList.getItems().add(question);
         }
         questionList.getSelectionModel().selectFirst();
-
-        // Aantal vragen in een quiz dynamisch laten tonen
-        aantalVragen.setText(String.format("Deze quiz bestaat uit %d vraag(en).",
-                questionDAO.getNumberOfQuestionsInAquiz(questionList.getSelectionModel().getSelectedItem().getIdQuiz())));
-        questionList.setOnMouseClicked(mouseEvent -> aantalVragen.setText(String.format("Deze quiz bestaat" +
-                " uit %d vraag(en).", questionDAO.getNumberOfQuestionsInAquiz(questionList.getSelectionModel().getSelectedItem().getIdQuiz()))));
-        questionList.setOnKeyPressed(keyEvent -> aantalVragen.setText(String.format("Deze quiz bestaat" +
-                " uit %d vraag(en).", questionDAO.getNumberOfQuestionsInAquiz(questionList.getSelectionModel().getSelectedItem().getIdQuiz()))));
-
         if (alleVragen.isEmpty()) {
             Alert alert = new Alert(Alert.AlertType.INFORMATION);
             alert.setContentText("Er zijn geen door jou aangemaakte vragen in de database. " +
                     "\nEr moeten eerst quizzen aangemaakt worden");
             alert.showAndWait();
             Main.getSceneManager().showWelcomeScene();
+        } else {
+            aantalVragen.setText(String.format("Deze quiz bestaat uit %d vraag(en).",
+                    questionDAO.getNumberOfQuestionsInAquiz(questionList.getSelectionModel().getSelectedItem().getIdQuiz())));
+            questionList.setOnMouseClicked(mouseEvent -> aantalVragen.setText(String.format("Deze quiz bestaat" +
+                    " uit %d vraag(en).", questionDAO.getNumberOfQuestionsInAquiz(questionList.getSelectionModel().getSelectedItem().getIdQuiz()))));
+            questionList.setOnKeyPressed(keyEvent -> aantalVragen.setText(String.format("Deze quiz bestaat" +
+                    " uit %d vraag(en).", questionDAO.getNumberOfQuestionsInAquiz(questionList.getSelectionModel().getSelectedItem().getIdQuiz()))));
         }
     }
 
@@ -68,10 +68,20 @@ public class ManageQuestionsController {
 
     public void doDeleteQuestion(){
         Question vraagOmTeVerwijderen = questionList.getSelectionModel().getSelectedItem();
-        questionDAO.deleteOne(vraagOmTeVerwijderen);
-        questionList.getItems().remove(vraagOmTeVerwijderen);
-        Alert bevestigVerwijderen = new Alert(Alert.AlertType.WARNING);
-        bevestigVerwijderen.setContentText("De vraag is succesvol verwijderd");
-        bevestigVerwijderen.show();
+        Alert vraagVerwijderen = new Alert(Alert.AlertType.CONFIRMATION);
+        vraagVerwijderen.setContentText("Weet je zeker dat je deze vraag wil verwijderen?");
+        vraagVerwijderen.setHeaderText("Vraag verwijderen");
+        vraagVerwijderen.setTitle("Bevestiging");
+        Optional<ButtonType> decision = vraagVerwijderen.showAndWait();
+        if (decision.get() == ButtonType.OK) {
+            questionDAO.deleteOne(vraagOmTeVerwijderen);
+            questionList.getItems().remove(vraagOmTeVerwijderen);
+            Alert bevestigVerwijderen = new Alert(Alert.AlertType.WARNING);
+            bevestigVerwijderen.setContentText("De vraag is succesvol verwijderd");
+            bevestigVerwijderen.show();
+        } else {
+            questionList.getItems().clear();
+            setup();
+        }
     }
 }
